pyproject.toml
[project]
name = "nlppp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "adapters>=1.2.0",
    "langchain-openai>=1.1.13",
    "langgraph>=1.1.6",
    "peft>=0.19.0",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.1",
    "requests>=2.33.1",
    "semanticscholar>=0.12.0",
    "sentence-transformers>=5.4.0",
]
